Update and rename keepalive.py to web.py
